// src/custom-fields/enums/field-type.enum.ts
export enum CustomFieldType {
  TEXT = 'text',
  TEXTAREA = 'textarea',
  NUMBER = 'number',
  DECIMAL = 'decimal',
  BOOLEAN = 'boolean',
  DATE = 'date',
  DATETIME = 'datetime',
  SELECT = 'select',
  MULTI_SELECT = 'multi_select',
  CURRENCY = 'currency',
  PERCENTAGE = 'percentage',
  EMAIL = 'email',
  PHONE = 'phone',
  FILE = 'file',
  USER = 'user',
  PROJECT = 'project',
  VENDOR = 'vendor',
}